refactor: 검증 메서드 이름 변경
